+++ src/@types/util.ts
@@ -1,0 +1,1 @@
+declare type PropsType<T> = T extends React.ComponentType<infer P> ? P : never;