+++ src/item/index.ts
@@ -1,2 +1,8 @@
-export { default as Popup, ApiPopup, Rarity } from './popup';
+export {
+  default as Popup,
+  PopupIntl,
+  ApiPopup,
+  ApiPopupIntl,
+  Rarity,
+} from './popup';
 export { default as FrameType } from './FrameType';