+++ src/item/popup/ApiBody/Corrupted.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { FormattedMessage } from 'react-intl';
 
 export interface Props {}
 
@@ -6,7 +7,12 @@ export default class Stats extends React.PureComponent<Props> {
   render() {
     return (
       <div className="unmet">
-        <span className="lc">Corrupted</span>
+        <span className="lc">
+          <FormattedMessage
+            id="poe.popup.corrupted"
+            defaultMessage="Corrupted"
+          />
+        </span>
       </div>
     );
   }