+++ src/item/popup/ItemPopup/index.ts
@@ -1,3 +1,0 @@
-import ItemPopup from './ItemPopup';
-export default ItemPopup;
-export { Rarity } from './props';