+++ src/item/popup/index.ts
@@ -1,5 +1,7 @@
-import Component from './ItemPopup';
+import Popup from './ItemPopup';
 
-export default Component;
+export default Popup;
+export { default as PopupIntl } from './ItemPopupIntl';
 export { default as ApiPopup } from './ApiPopup';
+export { default as ApiPopupIntl } from './ApiPopupIntl';
 export { Rarity } from './ItemPopup';