+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/eps1lon/poe-components-item/compare/v0.2.1...HEAD)
+### Added
+- `PopupIntl` and `ApiPopupIntl` for i18n support via `react-intl`. The original 
+  versions (`Popup` and `ApiPopup`) now have a `messages` prop which enables
+  i18n support. The `*Intl` components take their messages from an `IntlProvider`
+  component from `react-intl` in the component tree. 
+
 ## [0.2.1](https://github.com/eps1lon/poe-components-item/compare/v0.2.0...v0.2.1) (2018-06-12)
 ### Fixed
 - physical damage not showing